amenity_3layers fix
arcgis dataset -> remove hdb, cleaned the datasets with kml and naming issues: childcare_clean, parkfacilities, tourism_attractions, appended some with updated geojsons from onemap.

osm points of interests -> mapping is not as accurate due to low accuracy of long/lat or rather the decimal point is not super long so it maps to multiple locations. resolution: to map it to the nearest one and concat osm and onemap properties (reverse geolocated)